Namespace all API-related controllers
