+++ app/controllers/api_controller.rb
@@ -1,2 +1,2 @@
-class ApiController < ActionController::API
+class APIController < ActionController::API
 end
