+++ config/routes.rb
@@ -3,7 +3,7 @@ Rails.application.routes.draw do
 
   devise_for :teams
 
-  scope '/api' do
+  namespace 'api' do
     resources :standups, except: [:new, :edit] do
       resources :standup_team_members, path: 'team_members', except: [:new, :edit]
     end
